+++ client/src/components/UploadFile.jsx
@@ -1,0 +1,76 @@
+import React from "react";
+
+import { FileInput, Label } from "flowbite-react";
+import { useState } from "react";
+
+import axios from "axios";
+
+export default function UploadFile() {
+  const [extractedText, setExtractedText] = useState(null);
+
+  const handleFileUpload = async (event) => {
+    const file = event.target.files[0]; 
+
+    const formData = new FormData();
+    formData.append("file", file); 
+
+    try {
+      const response = await axios.post(
+        "http://localhost:3000/resumes",
+        formData
+      );
+
+      console.log(response.data.extractedText);
+      setExtractedText(response.data.extractedText); 
+    } catch (error) {
+      console.error("Error occurred:", error);
+    }
+  };
+
+  return (
+    <div
+      id="left-resume"
+      className="w-2/4 flex justify-center items-center border-r-2"
+    >
+      <div className="text-center">
+        <div className="flex w-full items-center justify-center">
+          <Label
+            htmlFor="dropzone-file"
+            className="dark:hover:bg-bray-800 flex h-64 w-full cursor-pointer flex-col items-center justify-center rounded-lg border-2 border-dashed border-gray-300 bg-gray-50 hover:bg-gray-100 dark:border-gray-600 dark:bg-gray-700 dark:hover:border-gray-500 dark:hover:bg-gray-600"
+          >
+            <div className="flex flex-col items-center justify-center pb-6 pt-5 px-4">
+              <svg
+                className="mb-4 h-8 w-8 text-gray-500 dark:text-gray-400"
+                aria-hidden="true"
+                xmlns="http://www.w3.org/2000/svg"
+                fill="none"
+                viewBox="0 0 20 16"
+              >
+                <path
+                  stroke="currentColor"
+                  strokeLinecap="round"
+                  strokeLineJoin="round"
+                  strokeWidth="2"
+                  d="M13 13h3a3 3 0 0 0 0-6h-.025A5.56 5.56 0 0 0 16 6.5 5.5 5.5 0 0 0 5.207 5.021C5.137 5.017 5.071 5 5 5a4 4 0 0 0 0 8h2.167M10 15V6m0 0L8 8m2-2 2 2"
+                />
+              </svg>
+              <p className="mb-2 text-sm text-gray-500 dark:text-gray-400">
+                <span className="font-semibold">Click to upload</span> or drag
+                and drop
+              </p>
+              <p className="text-xs text-gray-500 dark:text-gray-400">
+                PDF, DOCX, DOC, TXT
+              </p>
+            </div>
+            <FileInput
+              id="dropzone-file"
+              accept=".pdf,.docs,.doc"
+              className="hidden"
+              onChange={handleFileUpload}
+            />
+          </Label>
+        </div>
+      </div>
+    </div>
+  );
+}
